-add suggestions and do refactoring
